implement the new sorting algorithm for extensions, see the documentation near functions ext_cmp1() and ext_cmp().
All sorting decisions are now in one place so it is easy
to revise them.

NOTE
the major change is that now most specific patterns come first,
so there might be differences in how diaplans behave.
If you really really really need to revert to the old sorting order
while you adapt your dialplan, you can uncomment the '#if 0' line
in ext_cmp().



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@26216 f38db490-d61c-443f-a65b-d21fe96a405b