Merged revisions 43564 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r43564 | russell | 2006-09-24 10:58:10 -0400 (Sun, 24 Sep 2006) | 5 lines

Fix a CLI command registration issue where an erroneous message claiming that
"iax2 show provisioning" was already registered.  This was because this command
was registering itself as both the command, as well as the command it is
deprecating.  (issue #8022, reported by bjweeks, fixed by myself)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@43565 f38db490-d61c-443f-a65b-d21fe96a405b