Merged revisions 48472 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r48472 | file | 2006-12-14 12:36:12 -0500 (Thu, 14 Dec 2006) | 2 lines

Payload values on the RTP structure can change AFTER a bridge has started. This comes from the packet handling of the SIP response when indication that it was answered has been sent. Therefore we need to protect this data with a lock when we read/write. (issue #8232 reported by tgrman)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@48473 f38db490-d61c-443f-a65b-d21fe96a405b