Merged revisions 48955 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r48955 | russell | 2006-12-25 00:19:48 -0500 (Mon, 25 Dec 2006) | 6 lines

Fix an error introduced by copying and pasting the handling of the >= operator
for the MATH function.  If a single equal sign was used as an operator, the
function would treat it is as if it were the >= operator.  Now, it properly
handles it as an invalid operator.
(issue #8665, patch by tempest1)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@48956 f38db490-d61c-443f-a65b-d21fe96a405b