add comments that AC_USE_SYSTEM_EXTENSIONS and AST_PROG_LD do not work on FreeBSD - presumably they depend on some auto* feature that is not installed by default.
I am not sure on what is a proper fix. In my local copy
i simply comment them out.

The AST_PROG_LD is a long standing isse, there were attempts
to fix it in the past but probably not enough has been copied
to acinclude.m4, and i had forgotten about it because i
commented out this call in configure.ac long ago



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@51288 f38db490-d61c-443f-a65b-d21fe96a405b