As the comment in the diff says:
AST_INLINE_API() is a macro that takes a block of code as an argument.
Using preprocessor #directives in the argument is not supported by all
compilers, and it is a bit of an obfuscation anyways, so avoid it.
As a workaround, define a macro that produces either its argument
or nothing, and use that instead of #ifdef/#endif within the
argument to AST_INLINE_API().



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@51312 f38db490-d61c-443f-a65b-d21fe96a405b