Merged revisions 52645 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r52645 | russell | 2007-01-29 15:26:27 -0600 (Mon, 29 Jan 2007) | 6 lines

Fix a problem with packet-to-packet bridging and DTMF mode translation.  P2P
bridging can only be used when the DTMF modes don't match if the core is
monitoring DTMF in both directions.  Then, the core will handle the translation.
Otherwise, this bridging method can not be used.
(issue #8936)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@52646 f38db490-d61c-443f-a65b-d21fe96a405b