Merged revisions 53497 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r53497 | russell | 2007-02-07 17:52:45 -0600 (Wed, 07 Feb 2007) | 6 lines

When building libdb1.a, put the additional flags needed at the beginning of
ASTCFLAGS, instead of at the end.  This way, we ensure that we find the local
headers first before accidentally trying to use headers that exist in
locations specified in the ASTCFLAGS passed from the main Makefile.
(issue #8637, ovi)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@53498 f38db490-d61c-443f-a65b-d21fe96a405b