Merged revisions 55005 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r55005 | russell | 2007-02-16 16:48:22 -0600 (Fri, 16 Feb 2007) | 9 lines

Revert the change I did in revisions 54955, 54969, and 54970, in 1.2, 1.4, 
and trunk.  I decided that once a conference is created from meetme.conf,
it is acceptable behavior that the pin can not be changed until the
conference goes away.  I also added a note in meetme.conf to describe this
behavior.

We still have another issue in 1.4 and trunk where some conferences with no
users don't go away.  That is the real bug that needs to be addressed here.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@55006 f38db490-d61c-443f-a65b-d21fe96a405b