Merged revisions 70948 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r70948 | murf | 2007-06-21 16:29:50 -0600 (Thu, 21 Jun 2007) | 1 line

This little fix is in response to bug 10016, but may not cure it. The code is wrong, clearly. In a situation where you set the CDR's amaflags, and then ForkCDR, and then set the new CDR's amaflags to some other value, you will see that all CDRs have had their amaflags changed. This is not good. So I fixed it.
........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@70949 f38db490-d61c-443f-a65b-d21fe96a405b