Merged revisions 72805 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r72805 | russell | 2007-07-01 18:51:34 -0500 (Sun, 01 Jul 2007) | 5 lines

When appending lines to call files to keep track of retries, write a leading
newline just in case the original call file did not have a newline at the end.
This fix is in response to a problem I saw reported on the asterisk-users
mailing list.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@72806 f38db490-d61c-443f-a65b-d21fe96a405b