Merged revisions 76132 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r76132 | russell | 2007-07-20 13:22:24 -0500 (Fri, 20 Jul 2007) | 6 lines

Use the define that specifies the default length of an artificially created
DTMF digit in the ast_senddigit() function.  The define is set to 100ms by
default, which is the same thing that this function was using.  But, using
the define lets changes take effect in this case, as well as the others where
it was already used.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@76138 f38db490-d61c-443f-a65b-d21fe96a405b