Merged revisions 80820 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r80820 | russell | 2007-08-24 15:24:05 -0500 (Fri, 24 Aug 2007) | 7 lines

Improve the debouncing logic in the DTMF detector to fix some reliability
issues.  Previously, this code used a shift register of hits and non-hits.
However, if the start of the digit isn't clean, it is possible for the
leading edge detector to miss the digit.  These changes replace the flawed
shift register logic and also does the debouncing on the trailing edge as well.
(closes issue #10535, many thanks to softins for the patch)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@80821 f38db490-d61c-443f-a65b-d21fe96a405b