Merged revisions 81120 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r81120 | mmichelson | 2007-08-27 16:08:48 -0500 (Mon, 27 Aug 2007) | 7 lines

DTMF begin frames should be ignored so that when an agent acks a call with the '#' key,
he doesn't cause a queue's announce file to be interrupted. Also went ahead and did the
same for the '*' key and for ending a call.

(closes issue #10528, reported by deskhack, patched by me)


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@81121 f38db490-d61c-443f-a65b-d21fe96a405b