Merged revisions 81349 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r81349 | mmichelson | 2007-08-29 11:35:29 -0500 (Wed, 29 Aug 2007) | 12 lines

This patch, in essence, will correctly pause a realtime queue member and reflect those
changes in the realtime engine.

(issue #10424, reported by irroot, patch by me)

This patch creates a new function called update_realtime_member_field, which is a generic
function which will allow any one field of a realtime queue member to be updated. This patch
only uses this function to update the paused status of a queue member, but it lays the foundation
for persisting the state of a realtime member the same way that static members' state is maintained
when using the persistentmembers setting


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@81350 f38db490-d61c-443f-a65b-d21fe96a405b