Merged revisions 83589 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r83589 | murf | 2007-09-22 13:39:16 -0600 (Sat, 22 Sep 2007) | 1 line

This closes issue #10788 -- The exact same fixes are made here for the first arg in the for(arg1; arg2; arg3) {} statement, as were done for the 3rd arg. It can now be an assignment that will embedded in a Set() app, or a macro call, or an app call.
........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@83590 f38db490-d61c-443f-a65b-d21fe96a405b