access channel locks through ast_channel_lock/unlock/trylock and not through ast_mutex primitives.
To detect all occurrences, I have renamed the lock field in struct ast_channel
so it is clear that it shouldn't be used directly.

There are some uses in res/res_features.c (see details of the diff)
that are error prone as they try and lock two channels without
caring about the order (or without explaining why it is safe).



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@89293 f38db490-d61c-443f-a65b-d21fe96a405b