shuffle a little bit the content of header files to reduce dependencies. In this commit: - move the ast_register/unregister_app functions to module.h   to avoid the need to include pbx.h for the simpler apps; - move the ast_group structure to channel.h to remove the   dependency of app.h on linkedlists.h
Note, this is a long process that I am doing in small steps.

The main difficulty is that now for each subsystem we
have a single header (e.g. channel.h) included by the subsystem
provider (usually one file, e.g. channel.c) and by its clients
(dozens of them, e.g. we have some 70+ apps and 30+ functions).

This requires the clients to include all the extra headers
required by the provider (eg. lock.h, linkedlists.h, definitions
of substructures...) even though many of the clients would be
just happy with opaque struct declarations and function prototypes.

The long term plan is to eventually rectify this structure
so that the compilation can become faster, and also APIs
are more stable.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@89522 f38db490-d61c-443f-a65b-d21fe96a405b