Bring in video console support for chan_oss (and later chan_alsa too). This is disabled in the default build, you need to explicitly enable it compiling with
	make COPTS=-DHAVE_VIDEO_CONSOLE

In return, you will be able to do a video call with chan_oss, using
the webcam (or X11 grabbing) as local source, and rendering the
incoming stream on your screen. Currently supported formats are
h261, h263, h263+, h264, mpeg4 (all through the avcodec lib, part
of ffmpeg).

Incoming video is on the left, outgoing video is on the right,
while the center displays a keypad (if configured so).

Right clicking on the video windows increases the size,
center clicking reduces the size.
Dragging the mouse (with the left key) on the right window
while the X11 grabber is active moves the grab area.

This is the result of work by Sergio Fadda, Marta Carbone
and myself, all properly disclaimed to digium.

Note, there is a lot of work left to do in this module,
including adding support for Video4LinuxV2 (I have patches
from Matteo Brancaleoni which should be integrated),
and making the GUI a lot more friendly than it is now
(e.g. supporting merging or switching among multiple sources,
a text window, and more).



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@93144 f38db490-d61c-443f-a65b-d21fe96a405b