Merge changes from team/russell/codec_resample
This commit imports libresample for use in Asterisk.  It also adds a new codec
module, codec_resample.  This module uses libresample to re-sample signed linear
audio between 8 kHz and 16 kHz.

It also provides an alternative for converting between 16 kHz G.722 and 8 kHz
signed linear when using G.722, which will likely be useful as some people have
complained about volume issues when the current codec_g722 converts to 8 kHz 
signed linear.  But, to test this, you will have to disable the g722-to-slin and
g722-to-slin16 translators in codec_g722.c.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@95501 f38db490-d61c-443f-a65b-d21fe96a405b