Blocked revisions 111662 via svnmerge
........
r111662 | mmichelson | 2008-03-28 11:36:59 -0500 (Fri, 28 Mar 2008) | 9 lines

The copy_request function did not take into account the necessary null terminator
for the string to be copied into. This resulted in parse_request reading invalid
memory beyond the end of the string, and in some cases led to crashes. Thanks
to falves11 for providing the valgrind output which led to the closure of this issue.

(closes issue #12284)
Reported by: falves11


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@111663 f38db490-d61c-443f-a65b-d21fe96a405b