Merged revisions 112469 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r112469 | mmichelson | 2008-04-02 12:36:49 -0500 (Wed, 02 Apr 2008) | 21 lines

Merged revisions 112468 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r112468 | mmichelson | 2008-04-02 12:36:04 -0500 (Wed, 02 Apr 2008) | 13 lines

Fix a race condition in the manager. It is possible that a new manager event
could be appended during a brief time when the manager is not waiting for input.
If an event comes during this period, we need to set an indicator that there is an
event pending so that the manager doesn't attempt to wait forever for an event that
already happened.

(closes issue #12354)
Reported by: bamby
Patches:
      manager_race_condition.diff uploaded by bamby (license 430)
	  (comments added by me)


........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@112470 f38db490-d61c-443f-a65b-d21fe96a405b