Blocked revisions 114550 via svnmerge
........
r114550 | russell | 2008-04-22 16:14:55 -0500 (Tue, 22 Apr 2008) | 4 lines

I thought I was going to be able to leave 1.4 alone, but that was not the case.
I ran into some problems with G.722 in 1.4, so I have merged in all of the fixes
in this area that I have made in trunk/1.6.0, and things are happy again.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@114551 f38db490-d61c-443f-a65b-d21fe96a405b