Blocked revisions 118173 via svnmerge
........
r118173 | mmichelson | 2008-05-23 17:35:50 -0500 (Fri, 23 May 2008) | 15 lines

A new feature thanks to the fine folks at Switchvox!

If a deadlock is detected, then the typical lock information will be
printed along with a backtrace of the stack for the offending threads.
Use of this requires compiling with DETECT_DEADLOCKS and having glibc
installed.

Furthermore, issuing the "core show locks" CLI command will print the
normal lock information as well as a backtraces for each lock. This
requires that DEBUG_THREADS is enabled and that glibc is installed.

All the backtrace features may be disabled by running the configure
script with --without-execinfo as an argument


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@118174 f38db490-d61c-443f-a65b-d21fe96a405b