Update to the janitor project for making sure to be thread-safe when retrieving the value of a channel variable. This covers app_queue.
This commit also incorporates a logical change. Previously, if MixMonitor
is to be used to record the call, all the arguments were parsed first. Then
the MixMonitor app would be located. Now the order of these operations has
been swapped. Now the app is located first so that we only go through the
work of parsing the arguments if the app was found.

(closes issue #12742)
Reported by: snuffy
Patches:
      bug_12742.diff uploaded by snuffy (license 35)



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@118783 f38db490-d61c-443f-a65b-d21fe96a405b