Blocked revisions 125530 via svnmerge
........
r125530 | mmichelson | 2008-06-26 17:02:55 -0500 (Thu, 26 Jun 2008) | 10 lines

Backport of attended transfer queue_log patch from trunk.

This patch allows for attended transfers to be logged in the
queue_log the same way that blind transfers have always been.

It was decided by popular opinion on the asterisk-dev mailing
list that this should be backported to 1.4. Thanks to everyone
who gave an opinion.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@125531 f38db490-d61c-443f-a65b-d21fe96a405b