The monitor-join option for queues was deprecated in favor of using MixMonitor to mix audio. However, it was pointed out to me that because of this, the command set for the MONITOR_EXEC variable is ignored as well. This means that people can't do their own custom mixing commands at the end of recordings in order to make, for instance, stereo recordings of calls.
With this patch, app_queue will set the "joinfiles" variable for the channel's
monitor if MONITOR_EXEC is not zero-length. This means that for normal audio
mixing, MixMonitor is still the preferred choice, but we allow custom
mixing to be done with the two Monitor streams if desired.

(closes issue #12923)
Reported by: snyfer



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@125649 f38db490-d61c-443f-a65b-d21fe96a405b