Merged revisions 125649 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r125649 | mmichelson | 2008-06-26 19:15:54 -0500 (Thu, 26 Jun 2008) | 15 lines

The monitor-join option for queues was deprecated in favor of using
MixMonitor to mix audio. However, it was pointed out to me that because
of this, the command set for the MONITOR_EXEC variable is ignored as well.
This means that people can't do their own custom mixing commands at the end
of recordings in order to make, for instance, stereo recordings of calls.

With this patch, app_queue will set the "joinfiles" variable for the channel's
monitor if MONITOR_EXEC is not zero-length. This means that for normal audio
mixing, MixMonitor is still the preferred choice, but we allow custom
mixing to be done with the two Monitor streams if desired.

(closes issue #12923)
Reported by: snyfer


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@125650 f38db490-d61c-443f-a65b-d21fe96a405b