Blocked revisions 133575 via svnmerge
........
r133575 | russell | 2008-07-25 09:57:11 -0500 (Fri, 25 Jul 2008) | 8 lines

Modify the main page of the doxygen documentation to link to a new page dedicated
to Asterisk licensing information.  The licensing page includes the Asterisk license,
as well as a (not yet complete) list of 3rd party libraries that may be used, as well
as what license we receive them under.

Help filling out this list in the format that I have started in doxyref.h would be
much appreciated.  :)

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@133576 f38db490-d61c-443f-a65b-d21fe96a405b