Merging the imap_consistency_trunk branch to trunk.
For an explanation of what "imap_consistency" is,
please see svn revision 134223 to the 1.4 branch.

Coincidentally, this also fixes a recent bug report
regarding the inability to save messages to the new
folder when using IMAP storage since they will would
be flagged as "seen" and not be recognized as new
messages.


(closes issue #13234)
Reported by: jaroth




git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@136715 f38db490-d61c-443f-a65b-d21fe96a405b