Merged revisions 137780 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r137780 | seanbright | 2008-08-14 11:03:03 -0400 (Thu, 14 Aug 2008) | 8 lines

If we detect that we are no longer connected, try to reconnect a few times
before giving up.  This relies on the timeout settings in the freetds.conf
file and, unfortunately, on a recent version of FreeTDS (0.82 or newer).

I either need to change the current execs to be non-blocking (which I do
not want to do) or we have to force people to run with the latest and
greatest of FreeTDS.  I'm on the fence...

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@137781 f38db490-d61c-443f-a65b-d21fe96a405b