Blocked revisions 140057 via svnmerge
........
r140057 | murf | 2008-08-26 09:57:49 -0600 (Tue, 26 Aug 2008) | 21 lines

(closes issue #13366)
Reported by: erousseau

This was a reasonable enhancement request, which was
easy to implement. Since it's an enhancement, it 
could only be applied to trunk.

Basically, for accounting where "initiated" seconds
are billed for, if the microseconds field on the end
time is greater than the microseconds field for the
answer time, add one second to the billsec field.

The implementation was requested by erousseau, and
I've implemented it as requested. I've updated the
CHANGES, the cdr.conf.sample, and the .h files
accordingly, to accept and set a flag for the
corresponding new option. cdr.c adds in the extra
second based on the usec fields if the option is
set. Tested, seems to be working fine.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@140059 f38db490-d61c-443f-a65b-d21fe96a405b