Blocked revisions 144420 via svnmerge
I am pushing this up to trunk and beyond, because I am not
certain that this bug will happen in the same way in 
trunk and 1.6.x. If someone can document this problem
in 1.6.x or trunk, I can revert this block and push the
change up.

........
r144420 | murf | 2008-09-25 10:12:14 -0600 (Thu, 25 Sep 2008) | 25 lines

(closes issue #12101)
Reported by: MVF
Tested by: neutrino88, urzedo, murf, thiagofernandes

Many thanks to neutrino88 for this patch, which
solves a problem whereby channels get a CANCEL
request, respond to it properly, but end up 
in a hung state, infinitely being rescheduled.
This fix is a bit crude, in that catches the
problem at a rather late phase, but it may
prevent infinite rescheduling problems that
might still arise.

It might have been better to find out why,
in the course of protocol handling, the channel
was not destroyed, but we leave that to 
future generations.

Many thanks to urzedo and thiagofernandes for
their work in verifying that the patch code
indeed is being executing, and averting the
problem.



........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@144421 f38db490-d61c-443f-a65b-d21fe96a405b