Blocked revisions 144759 via svnmerge
................
r144759 | mmichelson | 2008-09-26 17:15:40 -0500 (Fri, 26 Sep 2008) | 14 lines

Blocked revisions 144758 via svnmerge

........
r144758 | mmichelson | 2008-09-26 17:14:59 -0500 (Fri, 26 Sep 2008) | 7 lines

This patch was applied to 1.4 but it completely
does not apply since the "found" pointer is not
passed in to this function. If this is going to
be backported, it needs to be done differently or
a deeper backport needs to be done.


........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@144761 f38db490-d61c-443f-a65b-d21fe96a405b