The logic surrounding the return value of ast_spawn_extension within ast_bridge_call was reversed.
This problem was observed when a blind transfer placed from
the callee channel of a test call failed.

While the problem I am solving here is exactly the same
as what was reported in issue #13584, the difference is
that this fix I am applying is trunk-only. Issue #13584
was reported against the 1.4 branch, and my tests
of 1.4's blind transfers appear to work fine.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@145553 f38db490-d61c-443f-a65b-d21fe96a405b