Blocked revisions 146640 via svnmerge
........
r146640 | mmichelson | 2008-10-06 10:29:56 -0500 (Mon, 06 Oct 2008) | 15 lines

This commit introduces a change to how the "joinempty"
and "leavewhenempty" options are configured in queues.conf.

Instead of using vague terms like "yes," "no," "loose," and
"strict," we now accept a comma-separated list of values
to determine when to consider a member available.

Extended details can be found in the queues.conf.sample
file. Note also that the above four referenced values are
still accepted for backwards-compatibility, but are mapped
internally to the new method of representing the option.

AST-105


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@146641 f38db490-d61c-443f-a65b-d21fe96a405b