Merged revisions 148376 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r148376 | mmichelson | 2008-10-10 16:21:45 -0500 (Fri, 10 Oct 2008) | 13 lines

The logic used when checking a peer got changed subtly
in the "kill the user" commit and caused calls relying
on the insecure setting to not work properly. I changed
for finding a peer back to how it was prior to that 
commit.

(closes issue #13644)
Reported by: pj
Patches:
      13644_trunkv2.patch uploaded by putnopvut (license 60)
Tested by: pj


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@148377 f38db490-d61c-443f-a65b-d21fe96a405b