Blocked revisions 148519 via svnmerge
........
r148519 | murf | 2008-10-13 11:14:38 -0600 (Mon, 13 Oct 2008) | 15 lines

Hmmm. Nobody (but me) is interested in seeing 
the trie info when they do 'dialplan show ...'
(even with debug set to non-zero); so I set up a 
   'dialplan debug [context]' cli command instead, 
to explicitly show just the trie info.  I even
added an extension_exists() call to make sure the
trie info is built. I moved the explanatory header
to above the extension loop to ensure it only prints
once. And it will do this now, whether debug is set
or not.

I removed the trie printing from the 'dialplan show' 
command entirely. 


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@148559 f38db490-d61c-443f-a65b-d21fe96a405b