Merged revisions 156167 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r156167 | mmichelson | 2008-11-12 11:38:33 -0600 (Wed, 12 Nov 2008) | 7 lines

When doing some tests, I was having a crash at the end of every call
if an attended transfer occurred during the call. I traced the cause to
the CDR on one of the channels being NULL. murf suggested a check in
the end bridge callback to be sure the CDR is non-NULL before proceeding,
so that's what I'm adding.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@156169 f38db490-d61c-443f-a65b-d21fe96a405b