Merged revisions 160703 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r160703 | murf | 2008-12-03 13:41:42 -0700 (Wed, 03 Dec 2008) | 11 lines

(closes issue #13597)
Reported by: john8675309
Patches:
      patch.13597 uploaded by murf (license 17)
Tested by: murf, john8675309

This patch causes the setcid func to update the CDR
clid after setting the channel field.

I also notice that in trunk, the num/number of 1.4 is
left out; I decided to include the option to use
either in trunk, so as not to have 1.4 upgraders
not to have problems.


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@160760 f38db490-d61c-443f-a65b-d21fe96a405b