Fix a callerid parsing issue. If someone formatted callerid like the following: "name <number>" (including the quotation marks), then the parts would be parsed as
name: "name
number: number

This is because the closing quotation mark was not discovered since the number
and everything after was parsed out of the string earlier. Now, there is a check
to see if the closing quote occurs after the number, so that we can know if we
should strip off the opening quote on the name.

Closes AST-158



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@160943 f38db490-d61c-443f-a65b-d21fe96a405b