Merged revisions 164675 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
................
r164675 | russell | 2008-12-16 10:00:29 -0600 (Tue, 16 Dec 2008) | 19 lines

Merged revisions 164672 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.4

........
r164672 | russell | 2008-12-16 09:56:37 -0600 (Tue, 16 Dec 2008) | 11 lines

Fix a memory leak related to the use of the "setvar" configuration option.

The problem was that these variables were being appended to the list of vars
on the sip_pvt every time a re-registration or re-subscription came in.
Since it's just a waste of memory to put them there unless the request was an
INVITE, then the fix is to check the request type before copying the vars.

(closes issue #14037)
Reported by: marvinek
Tested by: russell

........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.1@164677 f38db490-d61c-443f-a65b-d21fe96a405b