Merged revisions 169794 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r169794 | mmichelson | 2009-01-21 16:10:02 -0600 (Wed, 21 Jan 2009) | 17 lines

Fix a crash when saying certain numbers in Chinese

This commit fixes a crash that was occurring when attempting to
say a number between 10000 and 100000 due to dividing by 0.

This also removes some places where a "zero" is spoken when it
should not be.


(closes issue #14291)
Reported by: dant
Patches:
      say.c-14291.diff uploaded by dant (license 670)
Tested by: dant



........


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.6.0@169795 f38db490-d61c-443f-a65b-d21fe96a405b