Parking attempts made to one end of a bridge no longer will hang up due to a parking failure.
Parking attempts made using either one-touch, or doing either a blind or 
assisted transfer to the parking extension now keep up the bridge instead of
hanging up the attempted parked party. Normal causes for the parking attempt
to fail includes the specific specified extension (via PARKINGEXTEN) not being 
available or if all the parking spaces are currently in use. To avoid having
to reverse a masquerade park_space_reserve was made to provide foresight if
a parking attempt will succeed and if so reserve the parking space.

(closes issue #13494)
Reported by: mdu113

Reviewed by Russell: http://reviewboard.digium.com/r/133/



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@173211 f38db490-d61c-443f-a65b-d21fe96a405b