Do not assume that the bridge_cdr is still attached to the channel when the 'h' exten is finished executing.
It is possible for a masquerade operation to occur when the 'h' exten is operating. This operation moves
the CDR records around causing the bridge_cdr to no longer exist on the channel where it is expected to.
We can not safely modify it afterwards because of this, so don't even try.

(closes issue #14564)
Reported by: meric


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@179840 f38db490-d61c-443f-a65b-d21fe96a405b