Fix a memory leak in the ast_answer / __ast_answer API call.
For a channel that is not yet answered this API call will wait
until a voice frame is received on the channel before returning.
It does this by waiting for frames on the channel and reading them
in. The frames read in were not freed when they should have been.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@182171 f38db490-d61c-443f-a65b-d21fe96a405b