Randomize IAX2 encryption padding
The 16-32 byte random padding at the beginning of an encrypted IAX2 frame turns out to not be all that random at all.  This patch calls ast_random to fill the padding buffer with random data.  The padding is randomized at the beginning of every encrypted call and for every encrypted retransmit frame.

Review: http://reviewboard.digium.com/r/193/



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@182281 f38db490-d61c-443f-a65b-d21fe96a405b