audio_audiohook_write_list() does not correctly update sample size after ast_translate.
audio_audiohook_write_list() does not take into account that the sample size may change after translation depending on if the original frame is is 8khz or 16khz.  While no 16kz codecs are supported in 1.4 at the moment, this will save headaches in the future if they ever are.  the sample size is now updated after translating to reflect this possibility.  Thanks to jcolp and mmichelson for helping me work this out.

(issue AST-197)



git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@188287 f38db490-d61c-443f-a65b-d21fe96a405b