Prevent a crash when SIP blonde transferring an unbridged call.
If one attempts to use the attended transfer button on a SIP phone
to transfer an unbridged call (such as a call to an IVR) but hangs
up while the target of the transfer is still ringing, we need to not
crash.

The problem was that ast_hangup was called from outside the channel
thread.

AST-211



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@189097 f38db490-d61c-443f-a65b-d21fe96a405b