Ensure that by default only one console channel driver is loaded
This configuration file was changed to ensure that only one console channel driver
(chan_oss) is loaded by default, but the change would only work if chan_console
was not built. Now it will work as expected; if chan_alsa or chan_console are built
and installed, they will not be loaded unless explicity requested.



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@191955 f38db490-d61c-443f-a65b-d21fe96a405b