Fix a bug with setting t38pt_udptl at the user or peer level.
If an incoming call authenticated as a user or peer and t38pt_udptl was
not set to yes in general then no UDPTL session would be present and any
T38 related things would fail. This commit changes it so that if after
authenticating T38 is enabled but no UDPTL session is present one will be
created.

(issue AST-215)


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@192387 f38db490-d61c-443f-a65b-d21fe96a405b