Fix handling of the 'state_interface' option of the 'queue add member' CLI command.
This change relates to r184980, which was a backport of the state interface
changes to app_queue from trunk.  trunk and all of the 1.6.x branches are not
affected.

'queue add member' allows for specifying an interface to use for device state
when adding a queue member via CLI, but the validation code was not properly
updated to reflect this optional argument.

(closes issue #15198)
Reported by: loloski
Patches:
      05272009_app_queue.diff uploaded by seanbright (license 71)
Tested by: loloski


git-svn-id: http://svn.asterisk.org/svn/asterisk/branches/1.4@197024 f38db490-d61c-443f-a65b-d21fe96a405b